move episode and sweep logic out to their own thing
